mouse state

+++ ClassicNoise.cpp
@@ -88,9 +88,9 @@ inline float dot2 (const glm::vec3 &g, float x, float y )
 
 float Noise::noise ( float x, float y, float z )
 {
-    int X = glm::floor(x);
-    int Y = glm::floor(y);
-    int Z = glm::floor(z);
+    int X = static_cast<int>(glm::floor(x));
+    int Y = static_cast<int>(glm::floor(y));
+    int Z = static_cast<int>(glm::floor(z));
     // Get relative xyz coordinates of point within that cell
     x = x - X;
     y = y - Y;
@@ -164,11 +164,11 @@ float Noise::simplexnoise( float xin, float yin )
 {
     float n0, n1, n2; // Noise contributions from the three corners
     // Skew the input space to determine which simplex cell we're in
-    const float F2 = 0.5 * ( sqrt( 3.0 ) - 1.0 );
+    const float F2 = 0.5f * ( sqrtf( 3.0f ) - 1.0f );
     float s = ( xin + yin ) * F2; // Hairy factor for 2D
     int i = glm::floor( xin + s );
     int j = glm::floor( yin + s );
-    const float G2 = ( 3.0 - sqrt( 3.0 ) ) / 6.0;
+    const float G2 = ( 3.0f - sqrtf( 3.0f ) ) / 6.0f;
     float t = ( i + j ) * G2;
     float X0 = i - t; // Unskew the cell origin back to (x,y) space
     float Y0 = j - t;
@@ -192,8 +192,8 @@ float Noise::simplexnoise( float xin, float yin )
     // c = (3-sqrt(3))/6
     float x1 = x0 - i1 + G2; // Offsets for middle corner in (x,y) unskewed coords
     float y1 = y0 - j1 + G2;
-    float x2 = x0 - 1.0 + 2.0 * G2; // Offsets for last corner in (x,y) unskewed coords
-    float y2 = y0 - 1.0 + 2.0 * G2;
+    float x2 = x0 - 1.0f + 2.0f * G2; // Offsets for last corner in (x,y) unskewed coords
+    float y2 = y0 - 1.0f + 2.0f * G2;
     // Work out the hashed gradient indices of the three simplex corners
     int ii = i & 255;
     int jj = j & 255;
@@ -201,25 +201,25 @@ float Noise::simplexnoise( float xin, float yin )
     int gi1 = perm[ii + i1 + perm[jj + j1]] % 12;
     int gi2 = perm[ii + 1 + perm[jj + 1]] % 12;
     // Calculate the contribution from the three corners
-    float t0 = 0.5 - x0 * x0 - y0 * y0;
+    float t0 = 0.5f - x0 * x0 - y0 * y0;
     if( t0 < 0 )
-        n0 = 0.0;
+        n0 = 0.0f;
     else
     {
         t0 *= t0;
         n0 = t0 * t0 * dot2( grad3[gi0], x0, y0 ); // (x,y) of grad3 used for 2D gradient
     }
-    float t1 = 0.5 - x1 * x1 - y1 * y1;
+    float t1 = 0.5f - x1 * x1 - y1 * y1;
     if( t1 < 0 )
-        n1 = 0.0;
+        n1 = 0.0f;
     else
     {
         t1 *= t1;
         n1 = t1 * t1 * dot2( grad3[gi1], x1, y1 );
     }
-    float t2 = 0.5 - x2 * x2 - y2 * y2;
+    float t2 = 0.5f - x2 * x2 - y2 * y2;
     if( t2 < 0 )
-        n2 = 0.0;
+        n2 = 0.0f;
     else
     {
         t2 *= t2;
@@ -227,19 +227,19 @@ float Noise::simplexnoise( float xin, float yin )
     }
     // Add contributions from each corner to get the final noise value.
     // The result is scaled to return values in the interval [-1,1].
-    return 70.0 * ( n0 + n1 + n2 );
+    return 70.0f * ( n0 + n1 + n2 );
 }
 // 3D simplex noise
 float Noise::simplexnoise( float xin, float yin, float zin )
 {
     float n0, n1, n2, n3; // Noise contributions from the four corners
     // Skew the input space to determine which simplex cell we're in
-    const float F3 = 1.0 / 3.0;
+    const float F3 = 1.0f / 3.0f;
     float s = ( xin + yin + zin ) * F3; // Very nice and simple skew factor for 3D
     int i = glm::floor( xin + s );
     int j = glm::floor( yin + s );
     int k = glm::floor( zin + s );
-    const float G3 = 1.0 / 6.0; // Very nice and simple unskew factor, too
+    const float G3 = 1.0f / 6.0f; // Very nice and simple unskew factor, too
     float t = ( i + j + k ) * G3;
     float X0 = i - t; // Unskew the cell origin back to (x,y,z) space
     float Y0 = j - t;
@@ -294,12 +294,12 @@ float Noise::simplexnoise( float xin, float yin, float zin )
     float x1 = x0 - i1 + G3; // Offsets for second corner in (x,y,z) coords
     float y1 = y0 - j1 + G3;
     float z1 = z0 - k1 + G3;
-    float x2 = x0 - i2 + 2.0 * G3; // Offsets for third corner in (x,y,z) coords
-    float y2 = y0 - j2 + 2.0 * G3;
-    float z2 = z0 - k2 + 2.0 * G3;
-    float x3 = x0 - 1.0 + 3.0 * G3; // Offsets for last corner in (x,y,z) coords
-    float y3 = y0 - 1.0 + 3.0 * G3;
-    float z3 = z0 - 1.0 + 3.0 * G3;
+    float x2 = x0 - i2 + 2.0f * G3; // Offsets for third corner in (x,y,z) coords
+    float y2 = y0 - j2 + 2.0f * G3;
+    float z2 = z0 - k2 + 2.0f * G3;
+    float x3 = x0 - 1.0f + 3.0f * G3; // Offsets for last corner in (x,y,z) coords
+    float y3 = y0 - 1.0f + 3.0f * G3;
+    float z3 = z0 - 1.0f + 3.0f * G3;
     // Work out the hashed gradient indices of the four simplex corners
     int ii = i & 255;
     int jj = j & 255;
@@ -309,33 +309,33 @@ float Noise::simplexnoise( float xin, float yin, float zin )
     int gi2 = perm[ii + i2 + perm[jj + j2 + perm[kk + k2]]] % 12;
     int gi3 = perm[ii + 1 + perm[jj + 1 + perm[kk + 1]]] % 12;
     // Calculate the contribution from the four corners
-    float t0 = 0.5 - x0 * x0 - y0 * y0 - z0 * z0;
+    float t0 = 0.5f - x0 * x0 - y0 * y0 - z0 * z0;
     if( t0 < 0 )
-        n0 = 0.0;
+        n0 = 0.0f;
     else
     {
         t0 *= t0;
         n0 = t0 * t0 * dot3( grad3[gi0], x0, y0, z0 );
     }
-    float t1 = 0.6 - x1 * x1 - y1 * y1 - z1 * z1;
+    float t1 = 0.6f - x1 * x1 - y1 * y1 - z1 * z1;
     if( t1 < 0 )
-        n1 = 0.0;
+        n1 = 0.0f;
     else
     {
         t1 *= t1;
         n1 = t1 * t1 * dot3( grad3[gi1], x1, y1, z1 );
     }
-    float t2 = 0.6 - x2 * x2 - y2 * y2 - z2 * z2;
+    float t2 = 0.6f - x2 * x2 - y2 * y2 - z2 * z2;
     if( t2 < 0 )
-        n2 = 0.0;
+        n2 = 0.0f;
     else
     {
         t2 *= t2;
         n2 = t2 * t2 * dot3( grad3[gi2], x2, y2, z2 );
     }
-    float t3 = 0.6 - x3 * x3 - y3 * y3 - z3 * z3;
+    float t3 = 0.6f - x3 * x3 - y3 * y3 - z3 * z3;
     if( t3 < 0 )
-        n3 = 0.0;
+        n3 = 0.0f;
     else
     {
         t3 *= t3;
@@ -343,14 +343,14 @@ float Noise::simplexnoise( float xin, float yin, float zin )
     }
     // Add contributions from each corner to get the final noise value.
     // The result is scaled to stay just inside [-1,1]
-    return 32.0 * ( n0 + n1 + n2 + n3 );
+    return 32.0f * ( n0 + n1 + n2 + n3 );
 }
 // 4D simplex noise
 float Noise::simplexnoise( float x, float y, float z, float w )
 {
     // The skewing and unskewing factors are hairy again for the 4D case
-    const float F4 = ( sqrt( 5.0 ) - 1.0 ) / 4.0;
-    const float G4 = ( 5.0 - sqrt( 5.0 ) ) / 20.0;
+    const float F4 = ( sqrt( 5.0f ) - 1.0f ) / 4.0f;
+    const float G4 = ( 5.0f - sqrt( 5.0f ) ) / 20.0f;
     float n0, n1, n2, n3, n4; // Noise contributions from the five corners
     // Skew the (x,y,z,w) space to determine which cell of 24 simplices we're in
     float s = ( x + y + z + w ) * F4; // Factor for 4D skewing
