+++ experimental/voxelstructure.h
@@ -31,9 +31,9 @@ public:
 
     void fillnoise()
     {
-        for(int i = 0; i < blocks.size(); ++i)
-            for(int j = 0; j < blocks[0].size(); ++j)
-                for(int k = 0; k < blocks[0][0].size(); ++k)
+        for(size_t i = 0; i < blocks.size(); ++i)
+            for(size_t j = 0; j < blocks[0].size(); ++j)
+                for(size_t k = 0; k < blocks[0][0].size(); ++k)
                 {
                     blocks[i][j][k].data = glm::clamp(Noise::noise(i/12.f+4,j/12.f+4,k/12.f+4)*5,0.f,1.f)*255;
                 }
@@ -41,9 +41,9 @@ public:
 
     void fillsphere()
     {
-        for(int i = 0; i < blocks.size(); ++i)
-            for(int j = 0; j < blocks[0].size(); ++j)
-                for(int k = 0; k < blocks[0][0].size(); ++k)
+        for(size_t i = 0; i < blocks.size(); ++i)
+            for(size_t j = 0; j < blocks[0].size(); ++j)
+                for(size_t k = 0; k < blocks[0][0].size(); ++k)
                 {
                     blocks[i][j][k].data = pow(i-blocks.size()/2.f-1, 2) + pow(j-blocks.size()/2.f-1, 2) + pow(k-blocks.size()/2.f-1, 2) < pow((blocks.size()-1)/2.f, 2) ? 128 : 0;
                 }
